Switch names BurgerConstructor <-> BurgerIngredients
