+++ src/components/burger-ingredients/ingredient-items-container/ingredient-items-container.tsx
@@ -1,15 +1,15 @@
 import React from "react";
-import ConstructorItems from '../constructor-items/constructor-items';
+import IngredientItems from '../ingredient-items/ingredient-items';
 import PropTypes from 'prop-types';
 import {ingredientPropType, tabPropType} from '../../utils/prop-type';
 
-const constructorItemsContainer = (props:any) => {
+const IngredientItemsContainer = (props:any) => {
     const ingredients:any = {};
     props.tabs.forEach((tab: any) => ingredients[tab.type] = props.ingredients.filter((i:any) => i.type === tab.type));
 
     return (
         <>
-        { props.tabs.map((tab:any) => (<ConstructorItems 
+        { props.tabs.map((tab:any) => (<IngredientItems 
             title={tab.displayName} 
             key={tab.type}
             ingredients = { ingredients[tab.type] }
@@ -20,10 +20,10 @@ const constructorItemsContainer = (props:any) => {
     );
 }
 
-constructorItemsContainer.propTypes = {
+IngredientItemsContainer.propTypes = {
     ingredients: PropTypes.arrayOf(ingredientPropType).isRequired,
     tabs: PropTypes.arrayOf(tabPropType).isRequired,
     onSelectIngredient: PropTypes.func.isRequired
 }
 
-export default constructorItemsContainer
+export default IngredientItemsContainer
